Borsh fun-token no-locks metadata
