+++ examples/fun-token-old/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "fun-token"
+name = "fun-token-old"
 version = "0.1.0"
 authors = ["Near Inc <hello@nearprotocol.com>"]
 edition = "2018"
@@ -8,6 +8,8 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.45"
 near-bindgen = { path = "../../near-bindgen" }
 borsh = "0.5.0"
 wee_alloc = { version = "0.4.5", default-features = false, features = [] }
