+++ examples/fun-token-old/README.md
@@ -1,0 +1,10 @@
+# Fungible token
+
+Fungible token like ERC-20 but designed for the async runtime like NEAR.
+It includes locks and JSON serialization.
+
+## Testing
+To test run:
+```bash
+cargo test --package fun-token -- --nocapture
+```
